Implement suggestions from DCC-EX friends.
Set TMR2 period to 80us.

Implemented consist address feature.

Move variables out of dcc_decode and into globals for performance.

Moved checksum inside dcc_decode to make interface cleaner.

Reordered logic to check the speed/direction packets first, and the
function group packets second to improve performance.

Changed logic to not subtract checksum bit so my description of the number
of bytes in the message matches the NMRA docs.  Probably more comments to update
before this is complete.

Moved DCC forward and reverse to be a binary bit with constants.

Filled in a number of the TODOs with debug messages at least.

Updated motor control to handle NDOT correctly.