Moved CALCULATIONS to a spreadsheet.
Moved the calculations into a spreadsheet so that the values would
dynamically change when components needed to be updated.  Used the
dynamic capabilities along with the JLCPCB parts list to update
resistor values to have more common values and use JLCPCB basic
parts values.

Determined also that the calculations on the DCC input were wrong
in the previous file and added D21 to clamp the input and protect
the PIC.

Fixed the "Values" in the BOM so the components group properly.
This required also updating a number of silk screen items.

Added the copyright notice and URL to the back of the board.