Improved calculations, and other updates as a result.
Switched to tantalum capacitors per the data sheet recommendation
for the regulator.  Reworked board to accomodate.

Switched DCC input to be through a MOSFET to accomodate the track
voltage range.

Re-organized schematic so the blocks make more sense, changed to
US Letter paper, and re-annotated so each block has it's own 10's
range.

Updated CALCULATIONS.xlsx, including page formatting, reworked it
so there is one tab per block on the schematic, and re-checked all
calculations.  It now prints to paper pretty.

Updated all resistor values to match a 3.3v microprocessor and
JLCPCB basic parts availability.

Updated MCLR hold up circuit to include a diode to protect the rest
of the circuit while programming.

Moved the processor to the left.

Extended the board up 0.25" allowing the ICSP header to be moved
further away from the serial header.  This accomodates larger
programming adapters.

Moved the oscilloscope header up so the traces are shorter and the
board less crowded.