Correct over-voltage on high side drive.
The P-Channel MOSFET gate would be driven to 0v by the old driver
circuit.  When the DCC voltage is > 20v, this exceeds the Vgs
maximum of the part.

Added another resistor to make a voltage divider cutting the voltage
in half, and thus driving the gate at 15v when Vdrive is 30v.  Now
Vgs is always within range of the part.