Move DCC_IN and BackEMF to pins with appropriate capabilities.
STM32 does not allow software remapping of the periperhal input
pins.  The location in the PIC design (Breadboard-1) routes them
to pins on an STM32 that are designed for the oscellator, but can
be re-used.  As such they have no ADC or Timer triggers attached
to them.

DCC_IN was moved to port 20, which can be routed to any of
TIM1_CH2, TIM2_CH2, TIM3_CH1, TIM3_CH2, TIM1_CH3, TIM16_CH1N.

BackEMF was moved to pin 7, which can be routed to ADC_IN0.

The displaced P1C H-Bridge output was moved to pin 2, as it's more
than capable of outputting the necessary signal.