Initial partial port of software to the STM32 platform.
This code is destined to go with Breadboard-2, but for now is being
developed and tested on an NUCLEO-F103RB board.  It's done in
STM32CubeIDE and uses the full set of hardware abstrations.

The motor code and function code has not yet been ported, and a
sync between this and the PIC code needs to happen.